test(radio): migrate a11y
